backend/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
flake8==7.3.0
gunicorn==23.0.0
idna==3.11
iniconfig==2.3.0
mccabe==0.7.0
mindee==4.31.0
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pypdfium2==4.30.0
pytest==9.0.1
pytest-django==4.11.1
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
sqlparse==0.5.4
urllib3==2.6.1
whitenoise==6.11.0
